:recycle: Refactor: BaseEntity로 인한 Id삭제 및 코드 통일
BaseEntity Extends, 관계설정 마무리, 오타 수정

+++ backend/src/main/java/com/trippiece/backend/api/domain/repository/MarketRepository.java
@@ -2,5 +2,5 @@ package com.trippiece.backend.api.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MarketRepository extends JpaRepository<MarketRepository, Integer> {
+public interface MarketRepository extends JpaRepository<MarketRepository, Long> {
 }
